backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.2
pydantic-settings==2.7.0
python-dotenv==1.0.1
langchain==1.1.3
langgraph==1.0.3
langchain-openai==1.1.1
langchain-tavily==0.2.13
tavily-python==0.7.13
tavily==1.1.0
requests==2.32.3
slowapi==0.1.9